AgeGenderRecognitionApp/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
gast==0.2.2
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
image==1.5.32
importlib-metadata==1.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==2.0.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
